Refactor authentication to use Shared.Auth project
- Updated `CategoriesService.csproj` and `ReportingService.csproj` to reference `Shared.Auth` and `Shared.Messaging`.
- Replaced `System.Security.Claims` with `Shared.Auth.Extensions` in `Program.cs`.
- Refactored JWT authentication setup to use `AddStandardJwtAuthentication`.
- Added `ClientId` to `Jwt` section in `appsettings.json`.
- Included `Shared.Auth` project in the solution file.
- Created `JwtAuthenticationExtensions.cs` for JWT logic in `Shared.Auth`.
- Updated `TransactionsService.csproj` to reference `Shared.Auth` and modified its `appsettings.json` for authentication credentials.
