+++ ReportingService/ReportingService.csproj
@@ -32,4 +32,8 @@
     <Folder Include="Data\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Shared.Auth\Shared.Auth.csproj" />
+  </ItemGroup>
+
 </Project>
