+++ TransactionsService/appsettings.json
@@ -4,7 +4,8 @@
   },
   "Jwt": {
     "Authority": "http://localhost:8080/realms/adji",
-    "Audience": "account"
+    "Audience": "account",
+    "ClientId": "clear-budget"
   },
   "Auth": {
     "ClientId": "clear-budget",
